pyproject.toml
[project]
name = "simple-image-service"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "boto3>=1.38.31",
    "boto3-stubs==1.38.31",
    "fastapi>=0.115.12",
    "pillow>=11.2.1",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "ruff>=0.11.13",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
